fix(outline): fix outline effect
The outline effect cannot display properly and has been fixed.
